+++ src/serial/blocking.rs
@@ -1,7 +1,7 @@
 //! Blocking serial API
 
 /// Write half of a serial interface (blocking variant)
-pub trait Write<Word> {
+pub trait Write<Word = u8> {
     /// The type of error that can occur when writing
     type Error: crate::serial::Error;