+++ src/spi/blocking.rs
@@ -1,7 +1,7 @@
 //! Blocking SPI API
 
 /// Blocking transfer
-pub trait Transfer<W> {
+pub trait Transfer<W = u8> {
     /// Error type
     type Error: crate::spi::Error;
 
@@ -20,7 +20,7 @@ impl<T: Transfer<W>, W> Transfer<W> for &mut T {
 }
 
 /// Blocking write
-pub trait Write<W> {
+pub trait Write<W = u8> {
     /// Error type
     type Error: crate::spi::Error;
 
@@ -37,7 +37,7 @@ impl<T: Write<W>, W> Write<W> for &mut T {
 }
 
 /// Blocking write (iterator version)
-pub trait WriteIter<W> {
+pub trait WriteIter<W = u8> {
     /// Error type
     type Error: crate::spi::Error;
 
@@ -62,7 +62,7 @@ impl<T: WriteIter<W>, W> WriteIter<W> for &mut T {
 ///
 /// This allows composition of SPI operations into a single bus transaction
 #[derive(Debug, PartialEq)]
-pub enum Operation<'a, W: 'static> {
+pub enum Operation<'a, W: 'static = u8> {
     /// Write data from the provided buffer, discarding read data
     Write(&'a [W]),
     /// Write data out while reading data into the provided buffer
@@ -71,7 +71,7 @@ pub enum Operation<'a, W: 'static> {
 
 /// Transactional trait allows multiple actions to be executed
 /// as part of a single SPI transaction
-pub trait Transactional<W: 'static> {
+pub trait Transactional<W: 'static = u8> {
     /// Associated error type
     type Error: crate::spi::Error;